successfully pulling api data on location change in store
unsure if this is the right pattern to use. I want to keep logic in redux, not components, but redux-observable is difficult. I can't figure how to make the stream conditional, perhaps because I need to understand rxjs better. I also may want to mix redux-thunk with redux-observable, as they can solve different problems.