fetching protected and unprotected data from router epic!
this means you no longer have to declare fetches on componentDidMount. just declare what to fetch in your router epic, and it will pattern match the url. also, to improve this, you should pull in types from the router files, or perhaps they should be importing the router types in the store area? or perhaps, the router types in the store area should *be* the data used to generate the switch statements for react.